passwordless_pre/api/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy== 2.5.1
Flask-Cors==3.0.10
eciespy== 0.3.10
fastecdsa==2.1.5
Flask-HTTPAuth ==4.2.0
Werkzeug ==1.0.1
PyJWT ==2.0.1
requests ==2.25.1
umbral== 0.1.3a2
cryptography ==3.4.7
